Feature rest ap is

+++ user/src/main/java/com/stitch/user/repository/UserRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends RevisionRepository<UserEntity, Long, Long>, JpaRepository<UserEntity, Long>, JpaSpecificationExecutor<UserEntity> {
+public interface UserRepository extends RevisionRepository<UserEntity, Long, Integer>, JpaRepository<UserEntity, Long>, JpaSpecificationExecutor<UserEntity> {
     boolean existsByEmailAddress(String emailAddress);
 
     Optional<UserEntity> findByEmailAddress(String emailAddress);
